Finalizing renames and cleaning up import/build issues
