+++ CMakeLists.txt
@@ -3,9 +3,16 @@ project(ros_tcp_endpoint)
 
 find_package(catkin REQUIRED COMPONENTS
   rospy
+  std_msgs
+  message_generation
 )
 
 catkin_python_setup()
 
-catkin_package()
+add_message_files(DIRECTORY msg)
 
+add_service_files(DIRECTORY srv)
+
+generate_messages(DEPENDENCIES std_msgs)
+
+catkin_package(CATKIN_DEPENDS message_runtime)
